doc: separate the notion of a priority step and number
The notion of a priority steps was sometimes conflated with the notion
of priority number and sometimes not. It made the config documentation
really confusing.

I tried to make it clearer here. The priority step is strictly about the
notion of having steps of priorities: a first step when the most needed
segments are loaded, a second step when the second most nedded segments
are loaded, and so on. As such, steps are 1-indexed and only are here to
illustrate the concepts behind priorities which is that the segments are
downloaded by steps.

A priority number however is 0-indexed and is about the real value a
priority step will have in the code. Both `MAX_HIGH_PRIORITY_LEVEL` and
`MIN_CANCELABLE_PRIORITY` config values accept a priority number not a
priority step now.

I hope this made this complex part of the config (and of the RxPlayer)
clearer.