api/buffer: add `enableFastSwitching` loadVideo option
"Fast-switching" is a behavior allowing to replace low-quality
segments (i.e. with a low bitrate) with higher-quality segments
(higher bitrate) in the buffer.

This is a nice feature in most cases but in a few cases, the application
could want to disable it.

For example, some devices have poor support for segment replacement,
which is a necessary feature when fast-switching is enabled: Some
devices still decode the segments that should have been replaced, other
have decoding issues when a segment is replaced.

Because of that, I propose a new loadVideo option, `enableFastSwitching`
- which will be set to `true` by default - which will allow to enable /
  disable this optimization.