eme: [DO NOT MERGE] attach keystatuseschange event listener before load
This commit adds a `keystatuseschange` event listener before a
`mediaKeySession.load` call (to load a persisted MediaKeySession) and
removes it just after the load is done.

This is to check if the listener is called despite the fact that the
`keyStatuses` property is empty once the persisted MediaKeySession is
loaded.