Skip discontinuities at the end of Periods even when we are in it
This commit adds a conditon to skip discontinuities we previously
ignored.

Let's consider the following situation:
```
 = segment
 | Period boundary
 ^ current position
 - current checked range for segment download

|=======   --|====            |
 Period 1  ^   Period 2
```

Here we're at a position inside a "hole" at the end of the given
Period. Now let's consider that we found out that there's no segment to
load there.

In that situation, we previously just checked if there was a
discontinuity anounced in the Manifest. If not, we did nothing.
But we can also reliably seek to the next Period in any case, when the
Period is known to have ended.
This commit adds that supplementary logic.

As the logic in `check_for_discontinuity.ts` is very hard to follow I
also added some comments and removed an indentation level where I could.
This is the source of most of the diff.