update rxjs to version 7 beta10 and the RxPlayer source code accordingly
From a few test, it appears that RxJS 7 offers a slight performance
improvement over RxJS 6.

Any small gain is good to take on some resource-constrained set-top
boxes, this appears to save hundred of milliseconds (note that it's very
difficult to estimate) in some extreme cases.

However, RxJS is still in its beta phase. Yet judging by its last
changelogs, it appears stable enough, at least for non-production
environments for the moment.

RxJS changes several things, most of all in TypeScript typings which led
to several modifications linked to adding/removing types. No influence
on the resulting JS code for those.

But it also deprecated its `toPromise` method. This seems to be linked
to the fact that as opposed to Promise, an Observable can both:
  - emit multiple values
  - emit no value at all
Due to that, RxJS tells us to use either `lastValueFrom` or
`firstValueFrom` instead.

There is a key difference however with `toPromise`, in that you cannot
give it your own Promise implementation, which is something we do for
IE11 compatibility.

But as I found out a lot of releases ago, RxJS 6 already broke
compatibility at some point.

To use the RxPlayer on IE11, we're now forced to use Promise polyfills.
This is very sad and even technically a breaking change, but because it
was going for long without us noticing (due to the fact that our demo
page included those polyfills), it would make no sense to publish a
major version now for that.