Remove "progress" event from being listened by the clock
The clock, an internal mechanism running the main player logic on
different playback events (seek, loading, playback rate change, or at a
regular interval if none of those happen) was listening to the
`"progress"` event from the HTMLMediaElement.

This appears to be unnecessary, as `progress` doesn't seem to be an
important milestone enough to justify re-checking everything.
What this event is for on contents with a MediaSource is a little
blurry but it seems to be linked to when a new chunk is added.

On some CPU-constrained devices, each tick have been observed to take
between 5 and 30 ms. Progress events being one of the events the most
often emitted, this could consume in the hundred of milliseconds before
the content loads on those devices.