use shareReplay when initially loading the Manifest
The previous behavior just used a `share` which could lead to an issue
in the case the clean-up process took more time than fetching the
Manifest, we could have to download two times the Manifest.

Note that this has the inconvenient of keeping the event of the first
parsed Manifest in memory until the end of the content.