implement (again) parallel init and media segment fetching
This is yet another retry of #918 and #898 of loading when possible the
init segment parallelly to the first media segment(s).

We encountered a lot of issues when trying to implement that logic
before.
To simplify this implementation, I rely on the work to de-asynchronize
segment parsers done in #950, on which this PR is based.

The optimization should theoretically lead to lower loading time, as
both initialization and media segments are needed to play a content.

In reality, the browser might sometimes attenuate this with its internal
request re-scheduling and unrelated license fetching is most often the last
step before being able to play, so loading time advantages are still to
prove.

--

The implementation is done through a new class called the
`DownloadingQueue`.

The `RepresensationStream` gives the full list of segments it wants to
download to the `DownloadingQueue`, which performs
the whole request scheduling, sending events when segments are loaded.

This is a concept I long wanted to implement to simplify the interface
between the `Stream` and the `fetchers` (for now, the `fetchers` only
work single segment by single segment, not with the complete
sequential list).

For now, the `DownloadingQueue` has been implemented on the
`Stream`-side for simplicity reasons. However, I think we might want to
move that logic to the fetchers in future iterations, as I feel it would
be more in its place there.

--

I also long wondered if the initialization segment fetching operation
should only be parallelized with the first media segment request
(meaning than if the latter finished before, we do not perform new media
segment requests until the former is finished).
In the end, I chose to not have that limitation: the initialization and
first media segment will be loaded in parallel, then if the media
segment's download ends before the init, the next media segment will be
immediately requested also in parallel of the still-pending
initialization request.