Refacto initial seek and play logic
On some devices, mostly Samsung TVs (actually it's Samsung TVs and
another device for which the browser integration is not yet finished),
the potential initial seek (e.g. when playing live contents) was not
taken into account when we did it, which is when the `loaded-metadata`
event was received.

According to the WHATWG specification, what we did should be sufficient
but on those devices, the media element's `currentTime` property quickly
went back to `0`.

To fix that, we have found that a behavior worked every time: if we
seeked after the first audio and video segments have been pushed,
theseek is always considered.

However, I feel that this might not be optimal, it might be preferable
to seek as soon as it is possible, so the lower-level decoders could
technically prepare their buffers at the earliest possible time.

--

This commit implements a much more "brute-forcy" solution:
As soon as the "loaded-metadata" event arrive, we will try to seek at each clock
tick where the `currentTime` property is set to `0`.
We will do that until both `readyState` switches to at least `3` and the
`currentTime` is not at `0` anymore.

To implement that logic I found that I had to refacto the logic related
to the initial seek and play.

The main problem there is that `create_stream_clock.ts` relied on the
initial seek observable to know when it has to stop applying its
`wantedTimeOffset` to the position.
Not only that, but even in the current logic, there is a risk of a
potential future bug: `create_stream_clock.ts` relied on both the clock
and on that seek Observable which itself relied on the same clock.
If there were listened to in the wrong order (if
`create_stream_clock.ts` Observable was listen to before), there could
be subtle bugs arising.

This is the main reason why I merged the `initial_seek_and_play.ts` and
`create_stream_clock.ts` files, into a new `initial_seek_and_play.ts`
file.