Never stall on event "canplay" or "canplaythrough"
While debugging weird issues on Samsung TVs we noticed that
gaps in the media buffers considered as small enough where automatically
skipped by the browser.

Seeking those ourselves before the browser does it lead to huge decoding
problems because Samsung TVs had other peculiar behaviors.

So the solution we found was to detect when the browser will
automatically seek over a gap, and only perform the seek ourselves if it
does not.

After discussion with Samsung, I have made a theory: maybe we were
seeking the gap just before the browser was, due to a "canplay" or
"canplaythrough" event.
This commit, just do not consider the player as "stalled" when either of
those two events are received.
As those two events indicates that the browser should be able to play, it
makes sense. We would still have the next clock tick if that was not
actually the case.

This fixes the issues because the RxPlayer does not seek over gaps if
playback is not stalled.