EME: Fix possible leaks of MediaKeySessions if closed at the wrong time
This commit fixes a leak we only observed for now on some set-top-boxes
where the call to `generateRequest` is very long (around 1 second).

This leak concerned MediaKeySessions, which were all closed when
stoping a content due to the `closeSessionsOnStop` `keySystems` property
being set to `true`.
Here, when loading then stopping multiple contents rapidly, we could
observe some MediaKeySessions failing to be closed and thus can be
kept "alive" by the user-agent despite being completely removed from the
RxPlayer's point of view.

After investigation, it seems that we encountered the following case
in the [EME spec for the `close`
method](https://www.w3.org/TR/encrypted-media/#dom-mediakeysession-close):
`If this object's callable value is false, return a promise rejected
with an InvalidStateError.`

A `MediaKeySession` has an attached `callable` value, known only by the
user-agent (or at least it doesn't seem to be known as is on the
JS-side) which stays to false from the MediaKeySession's creation to
either the result of a `generateRequest` call or a `load` call.

Instead of putting in place a complex logic to only call `close` after
either of those two were done, I chose to try an initial `close` call
as soon as possible - like before - and handling the error when/if it
happened.

If this error seems to be due to aforementioned case (an
`InvalidStateError` on a MediaKeySession without a `sessionId`
property, we will retry after the sooner of those events:
  - a delay raising exponentially (for now starting at 100 ms and
    multiplied by 2 at each attemps)
  - after a `keystatuseschange` event (which indicates that the
    MediaKeySession should now be "callable", especially after a
    `load` call
 - after a `message` event, for the same reason than the previous
   point only this time it is more adapted when a `generateRequest`
   call is pending.

To avoid an infinite loop of retry, we will retry a maximum of 5 times.