improve container detection logic in the transport code
This commit fixes a minor bug and try improve how segment containers are
inferred in the transport code (in `src/transports`).

The initial issue was found while investigating #945, when I found out
that a plain-text media segment would be re-loaded indefinitely when the
`transportOptions.checkMediaSegmentIntegrity` `loadVideo` option is set
to `true` (thankfully this option should be rarely used and in the cases
where we know it is used, we do not depend on plain-text subtitles).

The source of that issue was that the segment integrity checks are being
run (among other conditions) if the function `isWEBMEmbeddedTrack`
returned `false`, assuming that a container not being a webm means that
is is an MP4.
I think that the reason behind that fallacy is that this function was
first written with only audio and video segments in mind, where that
assumption (at least for DASH contents) makes sense.

As plain-text subtitles are not WEBM, the MP4 checks run... and fail,
because it is not an MP4.

Also, as I dig deeper into what `isWEBMEmbeddedTrack` was doing, I noticed
that even that function assumed the segment being an audio or video
segment. That is also problematic as nothing in the name of the function
nor in the file's path indicate that this logic is restricted to those.

---

In the end, I decided to refacto (a very small one) how container
detection was handled in the transport code.

I added the `inferSegmentContainer` function, which infers a container
type of a segment of any type based on its mime-type and type of track.

It only can return 3 values:
  - `"mp4"`: when we know that segment is an MP4
  - `"webm"`: idem for webm
  - `undefined`: we cannot say for sure

For resilience reasons:
  - I authorized video to have an mp4 audio mimeType and vice-versa.
  - for audio and video, we consider that container is by default (when
    `inferSegmentContainer` returns `undefined`) and `MP4` file.
  - I do not try to detect a "plain-text" text track in
    `inferSegmentContainer`.
    This is assumed in the text track parser for when it is not clearly
    an MP4 nor a webm.

There's some improvements possible. Notably once the segment has been
loaded, we could also check the segment to see if it does seem to be an
MP4 or webm (either as a supplementary check, or when the current
`inferSegmentContainer` implentation returns `undefined`).
However I did not take the time to do that yet, feel free to do it if
you want it. That would only be used for resilience reasons and we may
find the opportunity in a later improvement on that code anyway.