transports: simplify segment parsers types
This commit proposes a simplification of types used by the segment
parsers by defining a unique segment data type instead of two: one for
the init segment and the other for media segments.

As `SegmentBuffers` - on which those segment data will be pushed - are
only generic based on a single segment data type, this is not really
necessary anyway.

The real reason behind creating two types was to better indicate that
text and image Representation had no initialization segment (by setting
their type to `null`) but considering that a parsed media segment could
theoretically be `null` as well, nothing is lost by using the same type
for both.

Being generic over a single type simplify the way we approach segment
parsers in my opinion. Now we're thinking: "which type is the parsed
segment data in?" not "which type is a parsed init segment data and
which type is a media segment data?". Both should generally be in the
same format and the only difference until now was that one could be
always `null`.
Also, considering that media segments could theorically contain
initialization data themselves, this is one more argument in the
direction of only defining a single type.

---

Another improvement in that commit is to limit `any` typings to the
small `segment_fetcher_creator.ts` file. Doing this there should reduce
possible typing errors as segment loader and parsers are now properly
type-checked (both their inputs and outputs).