Limit the postponment of a Manifest refresh to an acceptable value
Manifest files, in particular for live contents, may need to be
refreshed at a regular interval.

The logic behind scheduling Manifest refresh is actually complex.
This is because even if a "normal" refreshing delay is indicated in the
Manifest, the RxPlayer might prefer to postpone its update instead
because it judges that it is too frequent when compared with the time
it takes to parse it.

However, there were no higher limit to this postponment, which could
theorically lead to a very high delay before an update\*, and as such to
potential long buffering time (this could only happen if Manifest
parsing takes a lot of time though, but we could imagine some
exceptional conditions where this would be the case or even unrelated
reasons, like someone adding a breakpoint when inside the parsing logic).

This commit set a high limit to the postponment of this "automatic" refresh
logic. For now it cannot by default\** go higher than 6 times the
original calculated delay.

Even with this, there can be many other causes for Manifest updates (e.g.
the RxPlayer suspects that it is de-synchronized or an `emsg` box
declares that a new MPD should be obtained) which are not for now
impacted by this modification.

\* This is not linked to any bug we know of for now.
We did have some recent issue concerning the Manifest not refreshing,
but our investigation lead to a poor implementation of a `manifestLoader`
function done by the application instead.

\** The `minimumManifestUpdateInterval` `transportOptions` given to the
`loadVideo` API is not impacted by this and thus the delay can still be
higher.

---

I also found the `manifestUpdateScheduler` code to be difficult to
follow so decided to move some code around.