api: improve custom segmentLoader/manifestLoader integration
Before this commit, it was not possible to indicate that a Manifest or
segment request could be retried when using a custom `manifestLoader` or
`segmentLoader` (respectively for Manifest requests and segment
requests).

Consequently, the RxPlayer would always throw on the first error
encountered in either two of these APIs.

This was not too bad, as custom loader could "fallback" to our
implementation when it couldn't perform the request, and our
implementation does have a more advanced retry management.

Still, those APIs could be more powerful.

Consequently, I added two optional properties that can be set on a
rejected error (kind of like the `keySystems[].getLicense` API):

  - `canRetry`: indicate whether the RxPlayer __CAN__ retry the
    request. Even if `true`, the RxPlayer might decide no to based
    on its configuration.

  - `isOfflineError`: indicate whether the error is due to the user
    being offline (from the corresponding network).

    I added this one because the RxPlayer uses different retry counter
    depending on whether the error is due to an "offline error" or
    a ""regular"" error.

Those are both optional properties and it shouldn't change anything to
the previous behavior when they are not set.

As such, this commit shouldn't break the API.

I also rewrote much of the documentation for both APIs (`manifestLoader`
and `segmentLoader`) because it became too complex for just writing an
example code.