Remove CachedSegmentDetector from abr
The `CachedSegmentDetector` allowed to signal when a loaded segment had
a relatively high chance to come from the browser's cache.

It did so by checking if segments loaded abnormaly fast (for video
segments, this was under 50ms), and faster than other segments.

The aim was to ignore such segments when calculating the current bitrate.

--

However, we noticed that the ABR performance on some really
well-compressed and short contents was very poor when starting with the
lowest bitrate.

This was due to most of their segments being thought as loaded from the
cache and thus ignored. In the aforementioned case, not enough data was
in the end considered to make the first bitrate estimation.

--

For now, I choose to plainly remove the CachedSegmentDetector.

We might want to have a better approach in the future, such as noticing
segments loaded too fast relatively to the constated bitrate.