Add "freezing" notion and try to unfreeze
The issue
=========

This commit tries to resolve a very rare situation in which an infinite
rebuffering occurs for no apparent reason (media properties show nothing
explaining that situation).

When this happens, a very small seek seems to always unlock the
situation, making us think that the issue is lower-level than the player
and that forcing a buffer flush through seeking resets it.

This problem has however been reproduced very rarely:

  - some years ago, we reproduced frequently on some Multi-Period DASH
    contents which were ad-switching Canal+ testing streams

  - closer in time, I reproduced the same issue last month on Firefox.
    This error only occured once in around 15 `loadVideo` calls and
    I could not ever reproduce after restarting my machine

    I highly suspect a general package update (I'm talking about my
    linux distribution's applications, not npm) to be linked in some
    way, so this instability may be "normal"/rare enough here.

  - The most seen case was on Tizen, where it can randomly happens.
    I think it is linked to segments that are replacing other
    segments currently being played (a behavior that has since been
    greatly reduced due to those bad effects, even if it can still
    happen in very rare situations).

Solution
=========

Definition of `freezing` on the clock and removal of `stalled`
--------------------------------------------------------------

Previously, the RxPlayer's `clock` emitted a `stalled` property, which
was set in two occasions:
  - There was not enough data in the buffer
  - The `readyState` indicated that we could not play (readyState set to
    `1`).

Taking inspiration from Guillaume Renault's drafts on the subject, I chose
to split-up that property into two:

  - the `rebuffering` property, which is set when we are very close to
    buffer starvation.

    Like before in the same situation, even if this property is set, the
    video could still technically play, so it is still the job of the
    rest of the player's code to pause to facilitate buffer re-building
    when this happens.

  - the `freezing` property, which is set when we seem to be stuck at the
    same position despite having data in the buffer and not being paused
    (or ended or at a `playbackRate` of `0`).

As before, the `clock` only anounces those statuses and does not try to
resolve them. The resolution of both happens in separate modules.

A `stalled` event emitted by the `Init` module still exists, as the
`API` relies on it to emit the `BUFFERING`/`SEEKING` state. But now, this
event is derived from both the `rebuffering` and `freezing` clock
properties.

Both situations are also segregated in the `stalled` event by the usage
of different `reason` strings emitted with it. A freezing situation
leads to a `BUFFERING` state.

Separating the call to `play` and the `LOADED` state
----------------------------------------------------

Because this issue can prevent the LOADED state from being emitted, yet
detecting that situation without calling the somewhat linked `play`
method on the media element is difficult (if we're not playing, it is
normal to be stucked!), I had to separate the code doing the `play` call
from the code emitting the `LOADED` state.

The `play` call was previously only done when we could go right into the
LOADED state, which means a `readyState` of `3` or `4`.

Because the `readyState` has been seen to be set to `1` when the issue
happens, the `play` call is now performed when:
  - the [potential] seek is done
  - the `readyState` is set to `1`
  - we have enough data in the buffer to play
  - we are not still seeking

Which are less restrictive conditions than before (it was roughly the
same before, with the added condition of a `3` or `4` `readyState`).

As for the `LOADED` state, it still necessitate a `readyState` of ay
least `3`.

Limitations with this solution
==============================

This solution has still several limitations:

  - most importantly, false positives are a possibility.

    I see two main ones:

      1. On low-end devices, initial playback or seeking may be very
         slow to pick up. Here, we could think that we're inexplicably
         freezing, perform a small seek, and thus lengthen even more the
         time before playback.

         We may even prevent the media from playing by repeatedly doing
         that.

      2. When decryption keys are not yet obtained (e.g. the license
         request is still pending), we could be left in a situation
         where the RxPlayer think that is abnormally freezing.

    Moreover, we cannot rely on the user-agent for the first possibility
    (note: the shaka-player seems to rely on it for roughly the same
    issue) because the RxPlayer is present (and will be) on a multitude
    of very specific low-end devices.

    Adding an option would mean that we should document this very
    complicated case and guarantee the maintenance of that option's
    effect, which we're not sure about yet.

    For the second point, we could cross-reference with where we are in
    the EME negociations. However, doing so is very difficult and might
    not catch every cases either.

    For now I chose to """fix""" this by setting a sufficiently large
    delay before the seek is done - at least 6.5 seconds of freezing
    (could not decide between 6 or 7 :P) - and by seeking by a very
    small amount of time - 1 millisecond.

  - When the issue arises, we may still not emit a LOADED state when
`autoPlay` has not been set to `true`, because we cannot know if we're
"freezing" if we're not currently playing.