send available{Audio,Video}BitratesChange event for multi-Period contents
While browsing the code for a totally different thing, I noticed that
a condition in the private `_priv_triggerAvailableBitratesChangeEvent`
method in the API didn't make much sense.

Basically, we were only sending an event linked to a change in bitrate
choices if either:

  1. There was no bitrate choice previously (for example, because we
     just started the content).

  2. There was a bitrate choice previously, but it was the exact same
     choice that we have now.

The second choice doesn't make sense, but the reverse does (if the list
of available bitrates changed emit the event, if not don't).

This, plus the fact that I the bad habit of mixing-up my `if`s
conditions sometimes (programing is hard!) led me to believe that this
was an unfortunate typo.

---

Because this is very easy to test, I added some integration tests to
test not only that, but every events that should be linked to Period
changes.

Only one line in the code was updated, the huge majority of the diff
here are new integration test scenarios.