demo: fix disabling of the video track in the demo page
Audio-only in the demo page was broken due to a bad check on the video
Adaptation obtained through the `getAdaptations` API.

The case where the video `Adaptation` is `null` was not propertly
handled, leading to an error in the demo page, quickly after the video
is disabled.

I first tried to do it through optional chaining, but it appears that
there are many thing to configure so that eslint can parse thoses (we
have to set-up a babel config file somewhere, then use a "babel" parser
for eslint etc.).

I found it was not worth the hassle and just added a supplementary
condition.

This also could be shorte by checking for `!= null` instead of `!==
undefined` && `!== null` but I prefer to stay explicit here, not
everybody know that trick.