propose more helpful 'list' npm script instead of the huge output of 'info'
Replace old `info` npm script and its huge output with a less verbose
and more interactive `list` script`.

Today, running npm run `info` leads us to a huge output that is
difficult to navigate. Because the point of that script is to let
outsiders / new devs to discover at a single place what all available
scripts does, this is counter-productive.

Here, I tried to replace that script with a more subdivided and
interactive "list" script.