Add eslint "no-explicit-any" rule
This commit tries to heavily limit the usage of the `any` TypeScript
type, which is a type allowing to avoid type-checking for part of the
code where properly-defining / depending on the right type is difficult.

There was already few `any` usage in the RxPlayer's code as we've been
pretty careful:

  -  We added the `noImplicitAny` TypeScript compiler option (which
     forces us to realize when a type is considered any by being explicit
     about it).

  - We consider variables generally typed as `any`, such as thrown
    errors, as `unknown` instead.

    Unlike `any`, `unknown` is type-checked and thus needs to be
    properly infered/duck-typed before being used.

But we still were lazy in some situations, relying on `any` instead.

Those situations were:

  1. In tests, where we sometimes do weird non-type-safe things on
     purpose (such as when testing the API or browser-exposed code or
     when doing as if the browser misses or has other features).

  2. In `/compat`, as some browsers have JS-exposed API that either
     are not known or do not respect the TypeScript one.

  3. When writing in a type-safe matter was too difficult.

     Here I'm thinking mainly about the format of parsed segments, which
     is different depending on the type of data (text, audio/video,
     image...).

     At one end of the RxPlayer's logic those segment are loaded and
     parsed and at another end, they are pushed to the right
     `SegmentBuffer` implementation.
     In-between, we have type-agnostic code which was too difficult to
     properly define generically while still guaranteeing through typing
     that the right segment format went to the right SegmentBuffer.

     So `any` was used in some strategic points instead.

---

As relying on `any` is never ideal and can let bug pass through, I
decided to add through this commit, the
`@typescript-eslint/no-explicit-any` rule to eslint, our linter.

When `any` was used because guaranteeing type-safety was too difficult
in the RxPlayer's code (such as in the `SegmentBuffer`), the now
preferred method is to type it as `unknown` instead and essentially
perform type-checking at runtime (I still disabled this step in
production mode - when `__DEV__` is `false` to ensure we'll not trigger
regressions here).

In the vast majority of its usage where we purposely used any because of
testing code or of compatibility files, the now preferred solution is to
define ad-hoc types containing just what is needed and cast the variable
as it.

This is still unsafe, but much more safe than just relying on `any`.

---

I discovered and fixed two minor bugs in this PR, just by adding better
types:

  1. In the experimental `MediaCapabilitiesProber` tool, when calling
     `getCompatibleDRMConfigurations`, we could be left with `undefined`
     values in the returned array - behavior which does not correspond
     to the API documentation nor the type.

     For now, I just filter out all `undefined` from that array.
     However, I think that the developer forgot to write something here
     and I'm not 100% sure of what the original intent was. So there
     may be a supplementary issue here, to check.

  2. In `src/compat/eme/custom_media_keys/webkit_media_keys.ts`, which
     defines a custom WebKit implementation for EME APIs, we still
     called `someSession.close` even when we found out that
     the corresponding `someSession` was not defined.

     I don't know whether that create a real issue for application user
     but I doubt it. Still this is technically a bug as it was not at
     all the wanted behavior.