dash: Better handle $Time$-based low-latency contents
We've handled DASH-LL (a.k,a. LL-DASH), which are low latency DASH
contents, for more than 2 years.However, no application or packager we
worked with was ready in time to profit from that feature since then.

Fast-forward 2 years later, we found out that low-latency specs and
examples evolved a lot and were provided previously-never-seen
Time$-based low-latency contents.
The main issue with $Time$-based low-latency contents - which have a
`<SegmentTimeline>` XML element in the MPD - is that non-yet complete
segment could be anounced in the MPD.

Those "incomplete" segments would be generally announced with their
current unfinished duration, which may be much lower than the final
segment duration once it has been fully generated.
We did not have this problem with $Number $-based DASH-LL without a
`<SegmentTimeline>` element, as there no segment is individually
announced in the MPD.

This is problematic for the RxPlayer because the `duration` property
from a `ISegment` object, which is derived from a `<SegmentTimeline>`'s
`<S>` element when present, always implicitely meant the expected
**total** duration of the segment.
By subtly changing its meaning to "the currently known duration of
that segment" (and even **when** the `ISegment` object was generated
by the player), we risk to break some logic relying on the previous
definition.

So this commit tries to explicitly redefine the `duration` property and
also adds a supplementary `complete` boolean property to an `ISegment`:
  - When `complete` is set to `true` (most of the time), it means that
    the `duration` property truly reflects the duration of the full
    segment.
  - When `complete` is set to `false`, `duration` instead indicates the
    current and maybe incomplete duration of the full segment.

I then basically did a `grep duration` in the whole code and observed on
a case-by-case basis which usage was still valid and which needed
modification.
_(not as bad as it sounds as I had an idea of what to do, considering
that this commit is my second attempt at a $Time$-based DASH-LL solution,
the first one just being to set the `duration` property to `undefined`
for incomplete segments)_

I've also written an updated parsing logic for the`availabilityTimeOffset`
and `availabilityTimeComplete` attributes in the MPD, which are intimately
linked to DASH-LL.

There's still room for improvement, most notably the ABR logic could be
improved when loading incomplete segments, but I didn't do much on this
for now, considering @PaulRosset is already working on that subject
right now.