DASH: Handle ContentProtection Element present at the Representation level
The RxPlayer always considered the DASH-IF IOP specification first when
implementing DASH features.

DASH-IF IOP v4.3 indicates that a `ContentProtection` element should
always be in an `<AdaptationSet>` element and should apply to all
`<Representation>` it contains (DASH-IF IOP v4.3 chapter 7.7.2).

However, we encountered multiple MPD who put their `ContentProtection`
at the `<Representation>` level and some highly-used media players seems
to handle that case:
  - The ExoPlayer: https://github.com/google/ExoPlayer/blob/029a2b27cbdc27cf9d51d4a73ebeb503968849f6/library/dash/src/main/java/com/google/android/exoplayer2/source/dash/manifest/DashManifestParser.java#L712
  - The ShakaPlayer:
    https://github.com/google/shaka-player/blob/c993fca5793200a7d4ba556d6ef150db52365f68/lib/dash/dash_parser.js#L1168-L1173

(another library I looked at, dash.js, seems to only handle it at
AdaptationSet-level)

Also, the DASH specification (not the DASH-IF IOP) seems to authorize
this way of doing it.

Handling it at Representation-level both could make sense and is very
easy to handle in the RxPlayer (the parsed result is already part of our
internal `Representation` structure).

So this is only a proposal.

This commit just merge ContentProtection encountered at the `<AdaptationSet>`
level with the ones encountered at the `<Representation> level.