minor parsing errors when constructing IManifest object are now returned in a tuple
Previously, parsing errors obtained while parsing a
`Manifest`/`IManifest` object were set as a `contentWarnings` property
of that object.

This was kind of a hack, because constructor cannot return things
(excepted perhaps the constructed object?).
Now that the `Manifest` is not a class anymore but a TypeScript's
interface with functions creating it, it's now possible to have a
function returning both the Manifest object and warnings as separate
elements, which is arguably cleaner.

This is what this commit does.