Use MediaCapabilities API when creating an `IManifest` object
Previously, the RxPlayer only relied on the
`MediaSource.isTyepSupported` API to detect browser support, which only
does a check for codec support.

However, there's a newer set of APIs, the Media Capabilities APIs, which
allows to be more precize (by e.g. providing the resolution, framerate
etc.) and thus to detect if any of the characteristics of a content make
it non-decodable by the current platform.

Previously we did not rely on it both by fear of bringing regressions
(content which was playable becoming non-playable) and because codec
checks usually did the job.

However, we've seen a recent issue on Samsung TV that may be (we
couldn't test it for now sadly) detected by the Media Capabilities API:
4k contents led to a decoding error on 720p TVs.

After seeing what other players were doing, especially the shaka-player,
I decided to try this development which use the
`mediaCapabilities.decodingInfo` method to check support of video
contents.

This is kind of risky, and as such has to be tested on most targets, yet
this could fix current and future issues and the other information
returned by this API (about smoothness and power efficiency) may be very
useful for future improvements.

To test on 720p samsung TVs and other targets.