drm: Recreate MediaKeySession after unsuccesful MediaKeySession.load call
The `MediaKeySession.load(sessionId)` API is an EME (i.e.
encryption-related) API allowing to load a MediaKeySession persisted
by the browser. It returns a `Promise.<boolean>`.

From the EME recommendation, it seems like The `boolean` part indicates
if the given `sessionId` argument corresponds to the `sessionId` of a
MediaKeySession persisted by the browser.

In the current code, it was assumed that if it was set to `false`, then
our original MediaKeySession's state did not change much and as such, it
was safe to call `generateRequest` on it, so we can generate a challenge
for it instead.

However, recent errors make me believe that calling `generateRequest` on
such MediaKeySessions is not OK, as such MediaKeySessions might have
their `uninitialized` property (_non-visible from JS_) set to `false`,
which is a reason to throw in the `generateRequest` call as per the EME
spec.

This commit thus always recreate a persistent MediaKeySession from
scratch when `MediaKeySession.load` resolves with `false`.