DASH: RepresentationIndex' `getLastPosition` is now limited by the Period's end
The `getLastPosition` method of `RepresentationIndex`es was previously
ambiguous in that it returned the "end of the last segment in a given
Representation" but without precizing what happens if the parent
`Period` is shorter (in which case the segment is in the end cropped
before being played).

This commit sets a new - more precize - definition that changes a little
the behavior but should not break anything: it is now the theoretical
last playable position in the Representation.
As such it is now always bounded by the Period's end.

I did this to raise its usefulness: the last reachable position (the new
value) should be much more useful than the maximum position the last
segment contains media data for (the old value).
Moreover, we never really needed usage for the latter now, as the
Period's end was still relied on anyway.