[Proposal] Consider __RX_PLAYER_DEBUG_MODE__ boolean if declared in window
This change allows to simplify debugging sessions when either done
through [RxPaired](https://github.com/peaBerberian/RxPaired (once
https://github.com/peaBerberian/RxPaired/pull/1 is merged) or any other
tool by just setting `window.__RX_PLAYER_DEBUG_MODE__` to `true` before
the RxPlayer is evaluated.

The only influence this boolean has for now is to set the highest log
verbosity, a necessity to unlock all RxPaired features.
But we may add other features in the future.

This is not only restricted to the RxPaired tool by the way, any other
log-redirecting script could make use of that boolean.

The only risk I can think of is to have conflicting variables into
window but the name chosen seems rare enough to me for it being a real
problem.