Actually do not skip gap jumping when paused is `true`
This commit re-establish gap-jumping when the media element is paused as
the contrary broke functionalities, for example the possibility to
auto-skip a discontinuity at LOADING/RELOADING time.

In any case, only doing it when the application explicitely set the
`playbackRate` at `0` or at a negative number seems logical and
sufficient.