Replace autoTrackSwitch event by a more useful trackUpdate event
Until now, the RxPlayer v4.0.0 was scheduled to introduce the
`autoTrackSwitch` event to signal that the track of any type (audio,
video, text...) and Period had to be changed automatically by the RxPlayer
(for now the only - improbable - case is when the corresponding track
just disappears after a Manifest update - e.g. the corresponding
`AdaptationSet` element is removed from a DASH MPD).

This event was needed to perfectly replace the track preferences and
bitrate APIs - which are removed from the v4.0.0 for more powerful API -
as the track and wanted Representations would without this
event just silently be updated by the RxPlayer, without the application
knowing about it and being able to react upon it.

But after some tests, it seems that adding a much more general
`trackUpdate` event, signaling any type of track change for any type and
Period (that is, even when changed manually for example through a
`setAudioTrack` call) is much more easy-to-understand and useful.
And it isn't hard to implement.

Because an application might still want to know the reason for that
event (is it due to the track disappearing from the Manifest? It is
linked to a `setAudioTrack` call?), the `reason` property of that event
has been updated accordingly. It can now be set to:
  - `"manual"`: Explicit track modification, e.g. `setAudioTrack`
  - `"trickmode-enabled"`: Trickmode tracks have been enabled through
    `setPlaybackRate` (generally, this is linked to newly set video
    trickmode tracks)
  - `"trickmode-disabled"`: Trickmode tracks have been disabled through
    `setPlaybackRate` (generally, this is linked to newly unset video
    trickmode tracks).
  - `"missing"`: The track does not exist in the Manifest anymore

We still reserve the right to add new, non-colliding, reasons in the
future.