Do not merge "main" video AdaptationSet anymore
We previously merged multiple video AdaptationSet into a single video
track if they all had a Role set to 'main' due to a Canal+ need (which
is being able to adaptively switch between Representation in several
AdaptationSets due to protection constraints) and a
somewhat liberal interpretation of the following excerpt from the
DASH-IF IOP:

```
[...]
exactly one video Adaptation Set shall be signaled as the main one unless
different Adaptation Sets contain the same content with different quality
or different codecs. In the latter case, all Adaptation Sets with the
same content shall be signaled as the main content. Signaling as main
content shall be done by using the Role descriptor with
@schemeIdUri="urn:mpeg:dash:role:2011" and @value="main"
```

However, it turns out that there is both another, cleaner, way to
indicate that multiple AdaptationSet are "switchable" in the DASH-IF
IOP, as well as valid use-cases where you want different AdaptationSet
with the same content but not switchable between one another, for
example HDR-specific AdaptationSets.

By the way, I'm sure I've already made a PR on this in the past, but
couldn't find it.
Maybe it was cancelled due to some Canal+ needs? I have now the opposite
complaint so it should be good!